+++ app/build.gradle
@@ -63,8 +63,8 @@ dependencies {
       'androidx.lifecycle:lifecycle-runtime-ktx:2.2.0-alpha03',
       'com.google.dagger:dagger:2.24',
       "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version",
-      "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.2.1",
-      "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.2.1",
+      'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.2.1',
+      'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.2.1',
       'androidx.recyclerview:recyclerview:1.0.0',
   )
   testImplementation(
