+++ app/src/main/java/org/oppia/app/player/content/ContentCardAdapter.kt
@@ -1,11 +1,9 @@
 package org.oppia.app.player.content
 
 import android.content.Context
-import android.text.Html
 import android.text.Spannable
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.databinding.library.baseAdapters.BR
@@ -13,16 +11,16 @@ import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.content_card_items.view.*
 import org.oppia.app.R
 import org.oppia.app.databinding.ContentCardItemsBinding
-import org.oppia.app.databinding.LearnersCardItemBinding
+import org.oppia.app.databinding.InterationFeedbackCardItemBinding
 import org.oppia.data.backends.gae.model.GaeSubtitledHtml
-import org.oppia.util.data.UrlImageParser
+import org.oppia.util.data.HtmlParser
 
 const val CUSTOM_TAG = "oppia-noninteractive-image"
 const val HTML_TAG = "img"
 const val CUSTOM_ATTRIBUTE = "filepath-with-value"
 const val HTML_ATTRIBUTE = "src"
 const val VIEW_TYPE_CONTENT = 1
-const val VIEW_TYPE_INTERACTION = 2
+const val VIEW_TYPE_INTERACTION_FEEDBACK = 2
 
 /** Adapter to bind the contents to the [RecyclerView]. It handles rich-text content. */
 class ContentCardAdapter(
@@ -42,11 +40,11 @@ class ContentCardAdapter(
           DataBindingUtil.inflate<ContentCardItemsBinding>(inflater, R.layout.content_card_items, parent, false)
         ContentViewHolder(binding)
       }
-      VIEW_TYPE_INTERACTION -> {
+      VIEW_TYPE_INTERACTION_FEEDBACK -> {
         val inflater = LayoutInflater.from(parent.getContext())
         val binding =
-          DataBindingUtil.inflate<LearnersCardItemBinding>(inflater, R.layout.learners_card_item, parent, false)
-        LearnersViewHolder(binding)
+          DataBindingUtil.inflate<InterationFeedbackCardItemBinding>(inflater, R.layout.interation_feedback_card_item, parent, false)
+        InteractionFeedbackViewHolder(binding)
       }
       else -> throw IllegalArgumentException("Invalid view type")
     }
@@ -55,7 +53,7 @@ class ContentCardAdapter(
   override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
     when (holder.itemViewType) {
       VIEW_TYPE_CONTENT -> (holder as ContentViewHolder).bind(contentList!!.get(position).html)
-      VIEW_TYPE_INTERACTION -> (holder as LearnersViewHolder).bind(contentList!!.get(position).html)
+      VIEW_TYPE_INTERACTION_FEEDBACK -> (holder as InteractionFeedbackViewHolder).bind(contentList!!.get(position).html)
     }
   }
 
@@ -64,7 +62,7 @@ class ContentCardAdapter(
     return if (!contentList!!.get(position).contentId!!.contains("content") &&
       !contentList!!.get(position).contentId!!.contains(
         "Feedback")) {
-      VIEW_TYPE_INTERACTION
+      VIEW_TYPE_INTERACTION_FEEDBACK
     } else {
       VIEW_TYPE_CONTENT
     }
@@ -74,38 +72,21 @@ class ContentCardAdapter(
     return contentList!!.size
   }
 
-  inner class ContentViewHolder(val binding: ViewDataBinding) : RecyclerView.ViewHolder(binding.root) {
-    fun bind(rawString: String?) {
+ private inner class ContentViewHolder(val binding: ViewDataBinding) : RecyclerView.ViewHolder(binding.root) {
+   internal fun bind(rawString: String?) {
       binding.setVariable(BR.htmlContent, rawString)
       binding.executePendingBindings();
-      val html: Spannable = parseHtml(rawString, binding.root.tvContents)
-      binding.root.tvContents.text = html
+      val html: Spannable = HtmlParser(context,entity_type,entity_id).parseHtml(rawString, binding.root.tvContents)
+     binding.root.tvContents.text = html
     }
   }
 
-  inner class LearnersViewHolder(val binding: ViewDataBinding) : RecyclerView.ViewHolder(binding.root) {
-    fun bind(rawString: String?) {
+  private inner class InteractionFeedbackViewHolder(val binding: ViewDataBinding) : RecyclerView.ViewHolder(binding.root) {
+   internal fun bind(rawString: String?) {
       binding.setVariable(BR.htmlContent, rawString)
       binding.executePendingBindings();
-      val html: Spannable = parseHtml(rawString, binding.root.tvContents)
+      val html: Spannable = HtmlParser(context,entity_type,entity_id).parseHtml(rawString, binding.root.tvContents)
       binding.root.tvContents.text = html
     }
-  }
-
-   fun parseHtml(rawString: String?, tvContents: TextView): Spannable {
-    val html: Spannable
-    var htmlContent = rawString
-    if (htmlContent!!.contains(CUSTOM_TAG)) {
-      htmlContent = htmlContent.replace(CUSTOM_TAG, HTML_TAG, false);
-      htmlContent = htmlContent.replace(CUSTOM_ATTRIBUTE, HTML_ATTRIBUTE, false);
-      htmlContent = htmlContent.replace("&amp;quot;", "")
-    }
-    var imageGetter = UrlImageParser(tvContents, context, entity_type, entity_id)
-    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
-      html = Html.fromHtml(htmlContent, Html.FROM_HTML_MODE_LEGACY, imageGetter, null) as Spannable
-    } else {
-      html = Html.fromHtml(htmlContent, imageGetter, null) as Spannable
-    }
-    return html
   }
 }
